requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
